backend/requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
bcrypt==3.2.0
better-exceptions==0.3.3
blessed==1.17.12
boto3==1.17.25
botocore==1.20.27
cached-property==1.5.2
cachetools==4.2.1
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
cssutils==1.0.2
daphne==3.0.1
decorator==4.4.2
Django==3.1.7
django-cors-headers==3.7.0
django-countries==7.0
django-extensions==3.1.1
django-filter==2.4.0
django-managepy-anywhere==2.0
django-rest-knox==4.1.0
django-safedelete==1.0.0
django-seed==0.2.2
django-storages==1.11.1
djangorestframework==3.12.2
docker==4.4.3
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.14.1
emails==0.6
Faker @ file:///tmp/build/80754af9/faker_1615228226744/work
Faker==6.6.0
faker-enum==0.0.2
future==0.16.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
lxml==4.6.2
MarkupSafe==1.1.1
paramiko==2.7.2
pgclient==0.2.4
premailer==3.7.0
prompt-toolkit==3.0.17
psycopg2==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-dateutil==2.8.1
python-http-client==3.3.2
python-jose==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.3.4
semantic-version==2.5.0
sendgrid==6.6.0
service-identity==18.1.0
six @ file:///tmp/build/80754af9/six_1605205327372/work
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.3
Twisted==21.2.0
txaio==21.2.1
typing==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
websocket-client==0.57.0
zope.interface==5.2.0

numpy==1.20.2
pandas==1.2.3
scipy==1.6.2
scikit-learn==0.24.1
exrex==0.10.5
ipdb==0.13.6
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
Pygments==2.8.1
pygraphviz==1.7
toml==0.10.2
traitlets==5.0.5